refactor(atoms): merge SfCheckbox
