--- packages/vue/src/components/atoms/SfCheckbox/SfCheckbox.html
@@ -1,27 +1,0 @@
-<div
-  class="sf-checkbox"
-  :class="{'is-active': isChecked, 'is-disabled': disabled}"
->
-  <input
-    type="checkbox"
-    :id="name"
-    :name="name"
-    :value="value"
-    :checked="isChecked"
-    :disabled="disabled"
-    @input="inputHandler"
-    class="sf-checkbox__input"
-  />
-  <label :for="name" class="sf-checkbox__container">
-    <!-- @slot Custom check mark markup -->
-    <slot name="checkmark" v-bind="{isChecked, disabled}">
-      <div class="sf-checkbox__checkmark" :class="{'is-active': isChecked}">
-        <SfIcon v-if="isChecked" icon="check" size="11px" color="white" />
-      </div>
-    </slot>
-    <!-- @slot Custom label markup -->
-    <slot name="label" v-bind="{label, isChecked, disabled}">
-      <div class="sf-checkbox__label" v-if="label">{{label}}</div>
-    </slot>
-  </label>
-</div>
